+++ Changelog.md
@@ -11,14 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `inv_mod`, `gcd`, `gcd_extended`, `lcm`.
+- Added `inv_mod`, `mul_redc`, `gcd`, `gcd_extended`, `lcm`.
 - Added `sqlx` support.
+
+### Changed
+
+- Renamed `ring_inverse` to `inv_ring`.
 
 ## [1.2.0] — 2022-06-03
 
 ### Added
 
-- Added `reduce_mod`, `add_mod`, `mul_mod`, `pow_mod`, `inv_mod`.
+- Added `reduce_mod`, `add_mod`, `mul_mod`, `pow_mod`.
 - Added `num-bigint` and `ark-ff` support.
 - Made `algorithms` public, but with unstable API for now.